Stop the editor losing configuration it was asked to keep
Four ways a load-edit-save cycle destroyed something.

The properties panel cached one form page per widget *class*, so two widgets
of the same type shared a page. Selecting the second showed the first's
values, and Apply then wrote them into the second. The page is now rebuilt
per selection and discarded with it -- the form is read from the live config
on selection anyway, so there was nothing worth caching. That also stops
showUnsupported leaking a page into the stack every time.

Apply built its config from a default-constructed one, so any field the form
could not render was written back as the type default. It now starts from
the widget's live config and patches it, which is what the agent's
set_config path has always done.

Save hardcoded the window name to "editor_window", so every config it
touched came back renamed. The canvas now carries the name it loaded, the
way it already carries the background colour, and the panel's name field --
which was built but never read -- is wired to it.

Save also never checked the stream: writing to a path the process cannot
write reported "Saved config to ..." and dropped the layout.

Found while verifying the above, by diffing a config through a load/save
cycle: yaml-cpp encodes an 8-bit integer as a character, so the
speedometer's shift_box_markers: [28, 54, 87] was written to the file as
["\x1c", 6, W] -- a raw control byte, and a different list on the way back
in. The C125's center_page_digit had it too. Both are uint16_t now; nothing
in either needs to be one byte wide.

Two more in the same area: MainWindow tracked live widgets in a vector that
skipped ones that failed to build, while indexing into the config vector
that kept them, so after any failure a rebuild wrote over its neighbour --
each widget now carries its own config index. And the speedometer's marker
loop counted with a uint8_t, so 256 markers wrapped it and it never
terminated.

Verified by loading configs/dashboard/mercedes_190e_dash.yaml in the editor,
saving it, and diffing the parsed YAML: the only differences are two fields
written at the defaults the source omitted.

+++ dashboard/widgets/mercedes_190e_speedometer/include/mercedes_190e_speedometer/config.h
@@ -17,7 +17,11 @@ REFLECT_STRUCT(Mercedes190ESpeedometerConfig_t,
     (std::string, odometer_expression, ""),
     (std::string, odometer_zenoh_key, ""),
     (pub_sub::schema_type_t, odometer_schema_type, pub_sub::schema_type_t::VehicleOdometer),
-    (std::vector<uint8_t>, shift_box_markers, {})
+    // uint16_t, not uint8_t: yaml-cpp treats an 8-bit integer as a character,
+    // so `[28, 54, 87]` was emitted as `["\x1c", 6, W]` -- a raw control byte in
+    // the file and a different list on the way back in. Nothing here needs to be
+    // one byte wide, and a wider type keeps the YAML numeric.
+    (std::vector<uint16_t>, shift_box_markers, {})
 )
 
 #endif // MERCEDES_190E_SPEEDOMETER_CONFIG_H
